requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# docx_python.py: 3
docx_mailmerge == 0.4.0

# test.py: 1
gevent == 1.2.2

# docx_python.py: 7,8
python_docx == 0.8.6
